+++ libs/sdk-ui-ext/src/internal/interfaces/Visualization.ts
@@ -18,14 +18,6 @@ import {
     IExecutionConfig,
     LocalIdRef,
     ISortItem,
-    IAttributeSortTarget,
-    IAttributeSortType,
-    IMeasureSortTarget,
-    Identifier,
-    IAttributeLocatorItem,
-    newMeasureSort,
-    newAttributeSort,
-    newAttributeAreaSort,
 } from "@gooddata/sdk-model";
 import {
     ChartType,
@@ -40,6 +32,7 @@ import {
     SdkErrorType,
     VisualizationEnvironment,
 } from "@gooddata/sdk-ui";
+import { ISortConfig } from "./SortConfig";
 
 export type RenderFunction = (component: any, target: Element) => void;
 
@@ -322,94 +315,6 @@ export interface IVisualizationProperties {
     };
     [property: string]: any; // should not be used like this but to be backward compatible
 }
-
-export type AttributeSortSuggestion = {
-    type: "attributeSort";
-} & IAttributeSortTarget &
-    IAttributeSortType;
-
-export type MeasureSortSuggestion = {
-    type: "measureSort";
-} & IMeasureSortTarget;
-
-export type SortSuggestion = AttributeSortSuggestion | MeasureSortSuggestion;
-
-export function newMeasureSortSuggestion(
-    identifier: Identifier,
-    attributeLocators: IAttributeLocatorItem[] = [],
-): MeasureSortSuggestion {
-    const {
-        measureSortItem: { locators },
-    } = newMeasureSort(identifier, "asc", attributeLocators);
-    return {
-        type: "measureSort",
-        locators,
-    };
-}
-
-export function newAttributeSortSuggestion(identifier: Identifier): AttributeSortSuggestion {
-    const {
-        attributeSortItem: { attributeIdentifier },
-    } = newAttributeSort(identifier);
-    return {
-        type: "attributeSort",
-        attributeIdentifier,
-    };
-}
-
-export function newAttributeAreaSortSuggestion(
-    identifier: Identifier,
-    areaAggregation: "sum" = "sum",
-): AttributeSortSuggestion {
-    const {
-        attributeSortItem: { attributeIdentifier, aggregation },
-    } = newAttributeAreaSort(identifier, "asc", areaAggregation);
-    return {
-        type: "attributeSort",
-        attributeIdentifier,
-        aggregation,
-    };
-}
-
-/**
- * Specifies set of available sorts for given level of hierarchy:
- * Specific attribute to which sort is related - for eg. multiple ViewBy attributes, each can specify its sorting
- */
-export interface IAvailableSortsGroup {
-    // bucket item identifier
-    forBucketItem: LocalIdRef;
-    // all possible sort combinations for given level, sort direction is irrelevant
-    sortSuggestions: SortSuggestion[];
-    /**
-     * Additional text to available sorts, eg.
-     * when there is single available sort, this can contain explanation of the reason
-     */
-    explanation?: string;
-}
-export interface ISortConfig {
-    /**
-     * Current sort - default or chosen from inside of visualization
-     */
-    currentSort: ISortItem[];
-    /**
-     * All available sorts for current buckets
-     * - should contain current sort too
-     */
-    availableSorts: IAvailableSortsGroup[];
-    /**
-     * Whether sorting is supported by viz
-     */
-    supported: boolean;
-    /**
-     * Whether sorting is disabled for current buckets
-     */
-    disabled: boolean;
-    /**
-     * When sorting is disabled, this can contain explanation of reason
-     */
-    disabledExplanation?: string;
-}
-
 export interface IReferencePoint {
     buckets: IBucketOfFun[];
     filters: IFilters;
